+++ app/models/base.rb
@@ -2,5 +2,4 @@ class Base
   include Mongoid::Document
   include Mongoid::Timestamps
   include AccountScoped
-
 end
