+++ app/models/setup/connection.rb
@@ -1,25 +1,32 @@
 module Setup
   class Connection < Base
+    include NumberGenerator
+
+    has_many :webhooks, class_name: Setup::Webhook.name, inverse_of: :connection
+
+    devise :database_authenticatable
+
     field :name, type: String
     field :url, type: String
-    field :key, type: String
+    field :number, as: :key, type: String
     field :authentication_token, type: String
-    
-    devise :database_authenticatable, :token_authenticatable
 
-    has_many :webhooks, class_name: 'Setup::Webhook', inverse_of: :connection
-    
+    before_save :ensure_authentication_token
+
     accepts_nested_attributes_for :webhooks
 
     validates_presence_of :name, :url
     validates_uniqueness_of :authentication_token
-    
-    before_save :ensure_authentication_token
 
     def ensure_authentication_token
       self.authentication_token ||= generate_authentication_token
     end
-    
+
+    def generate_number(options = {})
+      options[:prefix] ||= 'C'
+      super(options) 
+    end
+
     rails_admin do 
       field :name
       field :url
@@ -30,14 +37,12 @@ module Setup
 
     private
 
-    def generate_authentication_token
-      loop do
-        token = Devise.friendly_token
-        break token unless User.where(authentication_token: token).first
+      def generate_authentication_token
+        loop do
+          token = Devise.friendly_token
+          break token unless User.where(authentication_token: token).first
+        end
       end
-    end
-    
- 
 
   end
 end
